develop: Move to more accessible OOP style, away from the long modular code
This was partly in development by the time the first few commits were started but
it was not finished. I decided to get back to this project to improve upon it as
there was not any proper documentation. I'll work on improving this with some
unit tests and potentially some more examples...

The biggest difference is that the lobby doesn't need nearly as much code as it did
as a module. Instead of a DIY framework I decided it was best to have functionality
that can be added or overwritten. This should help with getting started.

The new project has a basic chat lobby example as a nice example. Licensing has been
updated to GPL-v3.